[libromdata] KeyStoreUI::import3DSaeskeydb(): Fix casting warnings in 32-bit MSVC builds.
Output from MSVC 2010:
..\..\..\src\libromdata\crypto\KeyStoreUI.cpp(1219): warning C4244: 'initializing' : conversion from 'const int64_t' to 'unsigned int', possible loss of data
..\..\..\src\libromdata\crypto\KeyStoreUI.cpp(1220): warning C4244: 'argument' : conversion from 'const int64_t' to 'size_t', possible loss of data

+++ src/libromdata/crypto/KeyStoreUI.cpp
@@ -1216,8 +1216,8 @@ KeyStoreUI::ImportReturn KeyStoreUI::import3DSaeskeydb(const rp_char *filename)
 	}
 
 	// Read the entire file into memory.
-	unique_ptr<uint8_t[]> buf(new uint8_t[fileSize]);
-	size_t size = file->read(buf.get(), fileSize);
+	unique_ptr<uint8_t[]> buf(new uint8_t[(size_t)fileSize]);
+	size_t size = file->read(buf.get(), (size_t)fileSize);
 	if (size != (unsigned int)fileSize) {
 		// Read error.
 		// TODO: file->lastError()?
